+++ src/inet/linux/oshttpreq.c
@@ -11,8 +11,9 @@
 /* HTTP request (LibCURL-based implementation) */
 
 #include "../oshttpreq.inl"
-#include <core/arrst.h>
+#include <core/arrpt.h>
 #include <core/heap.h>
+#include <core/hfile.h>
 #include <core/stream.h>
 #include <core/strings.h>
 #include <sewer/cassert.h>
@@ -27,12 +28,23 @@ struct _oshttp_t
 {
     CURL *curl;
     struct curl_slist *headers;
-    String *host_url;
+    String *host;
     bool_t secure;
+    cookies_t cookies;
     Stream *resp_headers;
     Stream *resp_data;
     ierror_t error;
 };
+
+/*---------------------------------------------------------------------------*/
+
+#define COOKIE_DOMAIN 0
+#define COOKIE_SUBDOMAINS 1
+#define COOKIE_PATH 2
+#define COOKIE_SECURE 3
+#define COOKIE_TIMESTAMP 4
+#define COOKIE_NAME 5
+#define COOKIE_VALUE 6
 
 /*---------------------------------------------------------------------------*/
 
@@ -55,19 +67,16 @@ OSHttp *oshttp_create(const char_t *host, const uint16_t port, const bool_t secu
     OSHttp *http = heap_new(OSHttp);
     http->curl = curl_easy_init();
     http->error = ekIOK;
+    http->cookies = ekCOOKIES_ALL;
     http->secure = secure;
     http->headers = NULL;
     http->resp_headers = NULL;
     http->resp_data = NULL;
-
-    if (secure == TRUE)
-        http->host_url = str_printf("https://%s", host);
-    else
-        http->host_url = str_printf("http://%s", host);
+    http->host = str_c(host);
 
     if (http->curl != NULL)
     {
-        int res = curl_easy_setopt(http->curl, CURLOPT_PORT, port);
+        int res = curl_easy_setopt(http->curl, CURLOPT_PORT, (long)port);
         cassert_unref(res == CURLE_OK, res);
     }
     else
@@ -91,7 +100,7 @@ void oshttp_destroy(OSHttp **http)
         (*http)->headers = NULL;
     }
 
-    str_destroy(&(*http)->host_url);
+    str_destroy(&(*http)->host);
 
     if ((*http)->curl != NULL)
         curl_easy_cleanup((*http)->curl);
@@ -130,6 +139,14 @@ void oshttp_add_header(OSHttp *http, const char_t *name, const char_t *value)
 
 /*---------------------------------------------------------------------------*/
 
+void oshttp_cookies_policy(OSHttp *http, const cookies_t cookies)
+{
+    cassert_no_null(http);
+    http->cookies = cookies;
+}
+
+/*---------------------------------------------------------------------------*/
+
 static size_t i_write_response(char *buffer, size_t size, size_t nitems, void *userdata)
 {
     Stream *stm = (Stream *)userdata;
@@ -139,7 +156,7 @@ static size_t i_write_response(char *buffer, size_t size, size_t nitems, void *u
 
 /*---------------------------------------------------------------------------*/
 
-static void i_request(OSHttp *http, const bool_t use_get, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+static void i_request(OSHttp *http, const char *verb, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
 {
     int res = 0;
     cassert_no_null(http);
@@ -155,7 +172,7 @@ static void i_request(OSHttp *http, const bool_t use_get, const char_t *path, co
     cassert_unref(res == CURLE_OK, res);
 
     {
-        String *url = str_printf("%s%s", tc(http->host_url), path);
+        String *url = str_printf("%s://%s%s", http->secure ? "https" : "http", tc(http->host), path);
         res = curl_easy_setopt(http->curl, CURLOPT_URL, tc(url));
         cassert_unref(res == CURLE_OK, res);
         str_destroy(&url);
@@ -175,15 +192,34 @@ static void i_request(OSHttp *http, const bool_t use_get, const char_t *path, co
         cassert_unref(res == CURLE_OK, res);
     }
 
-    if (use_get == TRUE)
+    res = curl_easy_setopt(http->curl, CURLOPT_CUSTOMREQUEST, verb);
+    cassert_unref(res == CURLE_OK, res);
+
+    switch (http->cookies)
     {
-        res = curl_easy_setopt(http->curl, CURLOPT_HTTPGET, (long)1);
-        cassert_unref(res == CURLE_OK, res);
+    case ekCOOKIES_OFF:
+        /* Means, "don't read cookies" */
+        curl_easy_setopt(http->curl, CURLOPT_COOKIEFILE, NULL);
+        /* Means, "don't write cookies" */
+        curl_easy_setopt(http->curl, CURLOPT_COOKIEJAR, NULL);
+        /* Means, "remove in-memory cookies" */
+        curl_easy_setopt(http->curl, CURLOPT_COOKIELIST, "ALL");
+        break;
+
+    case ekCOOKIES_ALL:
+    {
+        String *cname = str_printf("%s_cookies.txt", tc(http->host));
+        String *cfile = hfile_appdata(tc(cname));
+        /* Means, "read cookies from this file" */
+        curl_easy_setopt(http->curl, CURLOPT_COOKIEFILE, tc(cfile));
+        /* Means, "write cookies to this file" */
+        curl_easy_setopt(http->curl, CURLOPT_COOKIEJAR, tc(cfile));
+        str_destroy(&cname);
+        str_destroy(&cfile);
+        break;
     }
-    else
-    {
-        res = curl_easy_setopt(http->curl, CURLOPT_POST, (long)1);
-        cassert_unref(res == CURLE_OK, res);
+    default:
+        cassert_default(http->cookies);
     }
 
     if (http->resp_headers != NULL)
@@ -221,14 +257,35 @@ static void i_request(OSHttp *http, const bool_t use_get, const char_t *path, co
 
 void oshttp_get(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
 {
-    i_request(http, TRUE, path, data, size, auto_redirect, error);
+    i_request(http, "GET", path, data, size, auto_redirect, error);
 }
 
 /*---------------------------------------------------------------------------*/
 
 void oshttp_post(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
 {
-    i_request(http, FALSE, path, data, size, auto_redirect, error);
+    i_request(http, "POST", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_put(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, "PUT", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_patch(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, "PATCH", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_delete(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, "DELETE", path, data, size, auto_redirect, error);
 }
 
 /*---------------------------------------------------------------------------*/
@@ -254,4 +311,116 @@ void oshttp_response_body(OSHttp *http, Stream *body, ierror_t *error)
     stm_write(body, data, size);
     stm_close(&http->resp_data);
     ptr_assign(error, ekIOK);
+}
+
+/*---------------------------------------------------------------------------*/
+/*
+* Cookie in Netscape format (\t as separator)
+* domain (eg: example.com or #HttpOnly_example.com)
+* include_subdomains TRUE or FALSE
+* path (eg: /)
+* secure TRUE if only HTTPS, FALSE if not
+* expiration Timestamp UNIX (seconds) or 0
+* name
+* value
+*/
+static bool_t i_process_netscape_cookie(char_t *data, const char_t **fields, const uint32_t size)
+{
+    uint32_t i = 0;
+    bool_t store = TRUE;
+    cassert_no_null(data);
+    cassert_no_null(fields);
+    while (*data != 0)
+    {
+        if (store == TRUE)
+        {
+            if (i < size)
+                fields[i] = data;
+            i += 1;
+            store = FALSE;
+        }
+
+        if (*data == '\t')
+        {
+            *data = 0;
+            store = TRUE;
+        }
+
+        data += 1;
+    }
+
+    return (bool_t)(i == 7);
+}
+
+/*---------------------------------------------------------------------------*/
+
+Stream *oshttp_cookies(OSHttp *http)
+{
+    Stream *stm = NULL;
+    cassert_no_null(http);
+
+    if (http->cookies == ekCOOKIES_ALL)
+    {
+        struct curl_slist *cookies = NULL;
+        struct curl_slist *nc = NULL;
+        bool_t comma = FALSE;
+        int res = 0;
+
+        res = curl_easy_getinfo(http->curl, CURLINFO_COOKIELIST, &cookies);
+        cassert_unref(res == CURLE_OK, res);
+
+        if (cookies != NULL)
+            stm = stm_memory(1024);
+
+        for (nc = cookies; nc != NULL; nc = nc->next)
+        {
+            const char_t *fields[7];
+            if (i_process_netscape_cookie(cast(nc->data, char_t), fields, sizeof(fields)) == TRUE)
+            {
+                if (comma == TRUE)
+                    stm_writef(stm, "; ");
+                else
+                    comma = TRUE;
+
+                stm_writef(stm, fields[COOKIE_NAME]);
+                stm_writef(stm, "=");
+                stm_writef(stm, fields[COOKIE_VALUE]);
+            }
+        }
+
+        curl_slist_free_all(cookies);
+    }
+
+    return stm;
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_cookie_delete(OSHttp *http, const char_t *name)
+{
+    struct curl_slist *cookies = NULL;
+    struct curl_slist *nc = NULL;
+    int res = 0;
+
+    cassert_no_null(http);
+    res = curl_easy_getinfo(http->curl, CURLINFO_COOKIELIST, &cookies);
+    cassert_unref(res == CURLE_OK, res);
+
+    for (nc = cookies; nc != NULL; nc = nc->next)
+    {
+        const char_t *fields[7];
+        if (i_process_netscape_cookie(cast(nc->data, char_t), fields, sizeof(fields)) == TRUE)
+        {
+            if (str_equ_c(fields[COOKIE_NAME], name) == TRUE)
+            {
+                String *cdel = str_printf("%s\t%s\t%s\t%s\t0\t%s\t", fields[COOKIE_DOMAIN], fields[COOKIE_SUBDOMAINS], fields[COOKIE_PATH], fields[COOKIE_SECURE], fields[COOKIE_NAME]);
+                res = curl_easy_setopt(http->curl, CURLOPT_COOKIELIST, tc(cdel));
+                cassert_unref(res == CURLE_OK, res);
+                str_destroy(&cdel);
+                break;
+            }
+        }
+    }
+
+    curl_slist_free_all(cookies);
 }