+++ src/inet/win/oshttpreq.c
@@ -11,6 +11,7 @@
 /* HTTP request (WinINet-based implementation) */
 
 #include "../oshttpreq.inl"
+#include "inet_win.inl"
 #include <core/arrst.h>
 #include <core/heap.h>
 #include <core/stream.h>
@@ -29,6 +30,9 @@
 #include <WinInet.h>
 #include <sewer/warn.hxx>
 
+#define INTERNET_SUPPRESS_COOKIE_PERSIST 0x00000003
+#define INTERNET_SUPPRESS_COOKIE_PERSIST_RESET 0x00000004
+
 struct _oshttp_t
 {
     HINTERNET hInternet;
@@ -36,6 +40,8 @@ struct _oshttp_t
     HINTERNET hRequest;
     ierror_t error;
     bool_t secure;
+    cookies_t cookies;
+    String *url;
     Stream *headers;
 };
 
@@ -59,7 +65,9 @@ OSHttp *oshttp_create(const char_t *host, const uint16_t port, const bool_t secu
 
     http->secure = secure;
     http->error = ekIOK;
+    http->cookies = ekCOOKIES_ALL;
     http->hInternet = InternetOpen(L"nappgui", INTERNET_OPEN_TYPE_DIRECT, NULL, NULL, 0);
+    http->url = str_printf("%s://%s", secure == TRUE ? "https" : "http", host);
     http->headers = stm_memory(1024);
     stm_set_write_utf(http->headers, ekUTF16);
 
@@ -96,8 +104,8 @@ void oshttp_destroy(OSHttp **http)
     if ((*http)->hInternet != NULL)
         InternetCloseHandle((*http)->hInternet);
 
+    str_destroy(&(*http)->url);
     stm_close(&(*http)->headers);
-
     heap_delete(http, OSHttp);
 }
 
@@ -127,6 +135,14 @@ void oshttp_add_header(OSHttp *http, const char_t *name, const char_t *value)
 
 /*---------------------------------------------------------------------------*/
 
+void oshttp_cookies_policy(OSHttp *http, const cookies_t cookies)
+{
+    cassert_no_null(http);
+    http->cookies = cookies;
+}
+
+/*---------------------------------------------------------------------------*/
+
 static void i_request(OSHttp *http, const WCHAR *verb, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
 {
     WCHAR wpath[1024];
@@ -152,6 +168,25 @@ static void i_request(OSHttp *http, const WCHAR *verb, const char_t *path, const
 
     flags |= http->secure ? INTERNET_FLAG_SECURE : 0;
     flags |= auto_redirect ? 0 : INTERNET_FLAG_NO_AUTO_REDIRECT;
+
+    switch (http->cookies)
+    {
+    case ekCOOKIES_OFF:
+        flags |= INTERNET_FLAG_NO_COOKIES | INTERNET_FLAG_RELOAD;
+        break;
+
+    case ekCOOKIES_ALL:
+    {
+        DWORD opt = INTERNET_SUPPRESS_COOKIE_PERSIST_RESET;
+        BOOL ok = InternetSetOption(http->hInternet, INTERNET_OPTION_SUPPRESS_BEHAVIOR, &opt, sizeof(DWORD));
+        cassert_unref(ok == TRUE, ok);
+        break;
+    }
+
+    default:
+        cassert_default(http->cookies);
+    }
+
     http->hRequest = HttpOpenRequest(http->hConnect, verb, wpath, L"HTTP/1.1", NULL, NULL, flags, 0);
     if (http->hRequest == NULL)
     {
@@ -194,6 +229,27 @@ void oshttp_get(OSHttp *http, const char_t *path, const byte_t *data, const uint
 void oshttp_post(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
 {
     i_request(http, L"POST", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_put(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, L"PUT", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_patch(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, L"PATCH", path, data, size, auto_redirect, error);
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_delete(OSHttp *http, const char_t *path, const byte_t *data, const uint32_t size, const bool_t auto_redirect, ierror_t *error)
+{
+    i_request(http, L"DELETE", path, data, size, auto_redirect, error);
 }
 
 /*---------------------------------------------------------------------------*/
@@ -272,4 +328,62 @@ void oshttp_response_body(OSHttp *http, Stream *body, ierror_t *error)
     }
 
     ptr_assign(error, ekIOK);
+}
+
+/*---------------------------------------------------------------------------*/
+
+Stream *oshttp_cookies(OSHttp *http)
+{
+    WString str;
+    Stream *stm = NULL;
+    const WCHAR *url = NULL;
+    TCHAR *buffer = NULL;
+    DWORD size = 0;
+    BOOL ok;
+    cassert_no_null(http);
+    url = wstring_init(tc(http->url), &str);
+    ok = InternetGetCookieEx(url, NULL, NULL, &size, INTERNET_COOKIE_HTTPONLY, NULL);
+    cassert_unref((ok == TRUE && size > 0) || (ok == FALSE && size == 0), ok);
+    /* Size in bytes (not in TCHARs) */
+    size /= sizeof(TCHAR);
+
+    if (size > 0)
+    {
+        buffer = heap_new_n(size, TCHAR);
+        /* Here size in TCHARs */
+        ok = InternetGetCookieEx(url, NULL, buffer, &size, INTERNET_COOKIE_HTTPONLY, NULL);
+        cassert_unref(ok == TRUE, ok);
+    }
+
+    if (buffer != NULL)
+    {
+        uint32_t csize = unicode_convers_nbytes(cast_const(buffer, char_t), ekUTF16, ekUTF8);
+        char_t *cbuffer = heap_new_n(csize, char_t);
+        unicode_convers(cast_const(buffer, char_t), cbuffer, ekUTF16, ekUTF8, csize);
+        stm = stm_memory(1024);
+        stm_writef(stm, cbuffer);
+        heap_delete_n(&buffer, size, TCHAR);
+        heap_delete_n(&cbuffer, csize, char_t);
+    }
+
+    wstring_remove(&str);
+    return stm;
+}
+
+/*---------------------------------------------------------------------------*/
+
+void oshttp_cookie_delete(OSHttp *http, const char_t *name)
+{
+    WString str1;
+    WString str2;
+    DWORD state;
+    const WCHAR *url = NULL;
+    const WCHAR *wname = NULL;
+    cassert_no_null(http);
+    url = wstring_init(tc(http->url), &str1);
+    wname = wstring_init(name, &str2);
+    state = InternetSetCookieEx(url, wname, L"deleted; expires=Thu, 01-Jan-1970 00:00:00 GMT", INTERNET_COOKIE_HTTPONLY, 0);
+    cassert_unref(state != FALSE, state);
+    wstring_remove(&str1);
+    wstring_remove(&str2);
 }