+++ src/osbs/unix/bthread.c
@@ -39,10 +39,22 @@ Thread *bthread_create_imp(uint32_t(func_thread_main)(void *), void *data)
 #pragma GCC diagnostic ignored "-Wcast-function-type"
 #endif
 #endif
+#if defined(__clang__)
+#if (__clang_major__ >= 16)
+#pragma clang diagnostic ignored "-Wcast-function-type-mismatch"
+#endif
+#endif
+
     int ret = pthread_create(thread, NULL, cast_func(func_thread_main, void *(*)(void *)), data);
+
 #if defined(__GNUC__)
 #if (__GNUC__ >= 8)
 #pragma GCC diagnostic warning "-Wcast-function-type"
+#endif
+#endif
+#if defined(__clang__)
+#if (__clang_major__ >= 16)
+#pragma clang diagnostic DEFAULT "-Wcast-function-type-mismatch"
 #endif
 #endif